refactor: rename deployment access tab folder
